+++ packages/akutil-component-template/test/index.js
@@ -1,11 +1,14 @@
-import { afterMutations } from 'akutil-common';
+import { afterMutations } from 'akutil-common-test';
 import { symbols } from 'skatejs';
 import chai from 'chai';
+import sinonChai from 'sinon-chai';
 import chaiAsPromised from 'chai-as-promised';
 import Component from '../src';
 
+chai.use(sinonChai);
 chai.use(chaiAsPromised);
 chai.should();
+
 const expect = chai.expect;
 const shadowRoot = symbols.shadowRoot;